shuffle backward compat functions into own file.
